prov/gni: Moved rbtree.* and fasthash.* outside of prov/gni.  Moved rbtree.* and fasthash.* from prov/gni/Makefile.am to Makefile.am.  Modified gnix_hashtable.c to refelct the fasthash changes.
