+++ Makefile.am
@@ -29,7 +29,9 @@ rdmainclude_HEADERS =
 # internal utility functions shared by in-tree providers:
 common_srcs = \
 	src/common.c \
-	src/enosys.c
+	src/enosys.c \
+	src/rbtree.c \
+	src/fasthash.c
 
 # ensure dl-built providers link back to libfabric
 linkback = $(top_builddir)/src/libfabric.la
@@ -48,6 +50,8 @@ src_libfabric_la_SOURCES = \
 	include/fi_list.h \
 	include/fi_signal.h \
 	include/fi_rbuf.h \
+	include/rbtree.h \
+	include/fasthash.h \
 	include/prov.h \
 	src/fabric.c \
 	src/fi_tostr.c \
