+++ include/fasthash.h
