--- prov/gni/common/fasthash/README
@@ -1,5 +1,0 @@
-fasthash.h and fasthash.c are the source code of FastHash.
-
-hashgen contains the framework source code used to derive FastHash. It
-also contains a dependency library which is located in the dep/
-directory.
