+++ src/fasthash.c
